create database bootstrap

+++ installer_scripts/js/initializeApp.js
@@ -1,8 +1,10 @@
 const fs = require("fs");
+const { resolve } = require("path");
 const { displayError, displayMessage } = require("./displayMessage.js");
 const { processExit } = require("./processExit.js");
 const { menu } = require("./menu.js");
 const { $, $$, $sh } = require("./shell.js");
+const { applyDatabaseConfig } = require("./applyDatabaseConfig.js");
 
 const DEBUG_DRY_RUN = false;
 
@@ -15,7 +17,7 @@ const pythonVersion = `3.10.11`;
 const pythonPackage = `python=${pythonVersion}`;
 const ffmpegPackage = `conda-forge::ffmpeg=4.4.2[build=lgpl*]`;
 const nodePackage = `conda-forge::nodejs=22.9.0`;
-const anacondaPostgresqlPackage = `anaconda::postgresql=12.17`;
+const anacondaPostgresqlPackage = `conda-forge::postgresql=16.4`;
 // const terraformPackage = `conda-forge::terraform=1.8.2`;
 const terraformPackage = ``;
 
@@ -27,9 +29,13 @@ const cudaChannels = [
 ].join(" -c ");
 const cpuChannels = ["", "pytorch"].join(" -c ");
 
+const windowsOnlyPackages =
+  process.platform === "win32" ? ["conda-forge::vswhere"] : [];
+
 const cudaPackages = `pytorch[version=${torchVersion},build=py3.10_cuda${cudaVersion}*] pytorch-cuda=${cudaVersion} torchvision torchaudio cuda-toolkit ninja`;
 const cudaPytorchInstall$ = [
   "conda install -y -k",
+  ...windowsOnlyPackages,
   terraformPackage,
   anacondaPostgresqlPackage,
   nodePackage,
@@ -41,6 +47,7 @@ const cudaPytorchInstall$ = [
 const cpuPackages = `pytorch=${torchVersion} torchvision torchaudio cpuonly`;
 const pytorchCPUInstall$ = [
   "conda install -y -k",
+  ...windowsOnlyPackages,
   terraformPackage,
   anacondaPostgresqlPackage,
   nodePackage,
@@ -114,13 +121,15 @@ Select the device (GPU/CPU) you are using to run the application:
   `
   );
 
-const gpuFile = "./installer_scripts/.gpu";
-const majorVersionFile = "./installer_scripts/.major_version";
-const pipPackagesFile = "./installer_scripts/.pip_packages";
+const getInstallerFilesPath = (file) => resolve(__dirname, "..", file);
+
+const gpuFile = getInstallerFilesPath(".gpu");
+const majorVersionFile = getInstallerFilesPath(".major_version");
+const pipPackagesFile = getInstallerFilesPath(".pip_packages");
 const majorVersion = "3";
 
 const versions = JSON.parse(
-  fs.readFileSync("./installer_scripts/versions.json")
+  fs.readFileSync(getInstallerFilesPath("versions.json"))
 );
 const newPipPackagesVersion = String(versions.pip_packages);
 
@@ -131,9 +140,7 @@ const readGeneric = (file) => {
   return -1;
 };
 
-const saveGeneric = (file, data) => {
-  fs.writeFileSync(file, data.toString());
-};
+const saveGeneric = (file, data) => fs.writeFileSync(file, data.toString());
 
 const readMajorVersion = () => readGeneric(majorVersionFile);
 const saveMajorVersion = (data) => saveGeneric(majorVersionFile, data);
@@ -143,9 +150,7 @@ const readGPUChoice = () => readGeneric(gpuFile);
 const saveGPUChoice = (data) => saveGeneric(gpuFile, data);
 
 const removeGPUChoice = () => {
-  if (fs.existsSync(gpuFile)) {
-    fs.unlinkSync(gpuFile);
-  }
+  if (fs.existsSync(gpuFile)) fs.unlinkSync(gpuFile);
 };
 
 const dry_run_flag = DEBUG_DRY_RUN ? "--dry-run " : "";
@@ -217,6 +222,11 @@ const FORCE_REINSTALL = process.env.FORCE_REINSTALL ? true : false;
 const initializeApp = async () => {
   displayMessage("Ensuring that python has the correct version...");
   await ensurePythonVersion();
+  try {
+    await applyDatabaseConfig();
+  } catch (error) {
+    displayError("Failed to apply database config");
+  }
   displayMessage("Checking if Torch is installed...");
   if (readMajorVersion() === majorVersion && !FORCE_REINSTALL) {
     if (await checkIfTorchInstalled()) {
@@ -292,17 +302,3 @@ function setupReactUI() {
   }
 }
 exports.setupReactUI = setupReactUI;
-
-async function setupDB() {
-  // initdb -D data/postgres -U postgres
-  // pg_ctl -D data/postgres -l data/postgres/logfile start
-  try {
-    displayMessage("Initializing database...");
-    await $sh("initdb -D data/postgres -U postgres");
-    // todo - manage DB process
-    // await $sh("pg_ctl -D data/postgres -l data/postgres/logfile start");
-    displayMessage("Successfully initialized database");
-  } catch (error) {
-    displayMessage("Failed to initialize database");
-  }
-}
